+++ lib/events/events.dart
@@ -2,11 +2,12 @@
 import 'package:event_bloc/event_bloc.dart';
 import 'package:flutter/material.dart';
 
-enum UIEvent<T> {
+enum VHCSiteEvent<T> {
   button<String>(),
   setTextScale<double>(),
   loadBlog<void>(),
   pickBlogCategory<String?>(),
+  changeBlogSearchTerm<String?>(),
   ;
 
   BlocEventType<T> get event => BlocEventType.fromObject(this);