+++ old_setup/Flutter/vhcsite/android/app/src/main/kotlin/com/vhcblade/vhcsite/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.vhcblade.vhcsite
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}