+++ old_setup/Flutter/vhcsite/lib/main.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:vhcsite/ui/main_app.dart';
-
-void main() {
-  runApp(const MainApp());
-}