+++ old_setup/Flutter/vhcsite/release_blog.sh
@@ -1,1 +1,0 @@
-dart script/release_blog.dart